Stop retaining messages in MQTT
There is literally no reason for retaining messages in the broker, in
fact, in the current scheme, it probably adds erroneous data, especially
if a message is not consumed quickly. The "N:<metric>" format that we
use to become collectd MQTT compliant means "NOW", which might very
well not be the case (it's not even really accurate, but it's fine for
our use case). If we consume such a message with a sufficient delay, the
data we record could be off by a substantial margin.
Switch retain to false to decrease this likelihood. This does mean that
we might see gaps in data until a new value is published, but given our
deep_sleep schedule, the effect of this on recent data is going to be a
smaller percentage than what is caused by deep_sleep

Signed-off-by: Alexandros Kosiaris <akosiaris@gmail.com>